Nest TopicExtractor to SimpleWriter
This class was design only for working with SimpleWriter, so nesting it makes
this clear.
